% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordflex.R
\docType{data}
\name{data_toy_mvordflex}
\alias{data_toy_mvordflex}
\title{Simulated toy data set}
\format{
A data frame with 1500 rows and 6 variables
}
\usage{
data(data_toy_mvordflex)
}
\description{
A data set containing two simulated ordinal responses with seven categories R1 and R2 a simulated
binary variable D1, seven quantitative covariates. The data is in a long format, where all the responses
have been stacked in one column. Additionally, the data set contains a column indicating
the subject id, year id and outcome id. In total, the three dimensional data set contains information
on 100 subjects over 5 years.

\itemize{
\item \code{firm_id} integer subject id
\item \code{year_id} integer year id
\item \code{outcome_id}  character id of the ordinal response ( "R1", "R2" or "D1")
\item \code{response} ordinal outcome containing all multiple measurements
\item \code{X1} continuous covariate
\item \code{X2} continuous covariate
}
}
