% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_struct.R, R/error_struct_MMO3.R
\name{error_struct}
\alias{error_struct}
\alias{cov_general}
\alias{cor_general}
\alias{cor_ar1}
\alias{cor_equi}
\alias{cor_MMO3}
\alias{cor_MMO3_cross}
\alias{cor_MMO3_ar1}
\title{Error Structures in mvord}
\usage{
cov_general(formula = ~1, value = numeric(0), fixed = FALSE)

cor_general(formula = ~1, value = numeric(0), fixed = FALSE)

cor_ar1(formula = ~1, value = numeric(0), fixed = FALSE)

cor_equi(formula = ~1, value = numeric(0), fixed = FALSE)

cor_MMO3(formula = ~1, value = numeric(0), fixed = FALSE)

cor_MMO3(formula = ~1, value = numeric(0), fixed = FALSE)

cor_MMO3_cross(formula = ~1, value = numeric(0), fixed = FALSE)

cor_MMO3_ar1(formula = ~1, value = numeric(0), fixed = FALSE)
}
\arguments{
\item{formula}{\code{\link{formula}} object}

\item{value}{specifies values of the correlation (and variance) parameters.
specifies values of the correlation (and variance) parameters. For \code{cor_MMO3()}
the first parameters correspond to the cross-sectional error structure while the
latter correspond to the AR(1) coefficients for each outcome. Default is \code{value
= numeric(0)} object. In this case the correlation parameters are initialized with zero.}

\item{fixed}{logical specifying whether the parameters of the error structure should not be optimized in the procedure, but will be \cr
fixed to the values specified in the argument \code{value}. Defaults to \code{fixed = FALSE}.}
}
\description{
Different \code{error.structures} are available in \pkg{mvord}:
\itemize{
\item general correlation structure (default) \code{cor_general(~ 1)},
\item general covariance structure \code{cov_general(~ 1)},
\item factor dependent correlation structure \code{cor_general(~ f)},
\item factor dependent covariance structure \code{cov_general(~ f)},
\item equicorrelation structure \code{cor_equi(~ 1)},
\item covariate dependent equicorrelation structure \code{cor_equi(~ S)},
\item AR(1) correlation structure \code{cor_ar1(~ 1)}, or
\item covariate dependent AR(1) correlation structure \code{cor_ar1(~ S)}.
}
For more details see vignette.

Different \code{error.structures} are available in \pkg{mvordflex}.
For three dimensional data in particular we have
\itemize{
\item correlation structure based on a multivariate AR(1) process \code{cor_MMO3(~ 1)}
\item correlation structure for three dimensional panel data where only
      cross-sectional correlation is accounted for \code{cor_MMO3_cross(~ 1)}
\item correlation structure for three dimensional panel data where only longitudinal
      correlation is accounted for by an AR(1) structure \code{cor_MMO3_ar1(~ 1)}
}
}
