% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvordflex.R
\name{mvordflex-package}
\alias{mvordflex-package}
\title{Multivariate Ordinal Regression Models - Extension to Three-Dimensional Panel Data.}
\description{
The R package mvordflex implements composite likelihood estimation in the class of multivariate
ordinal regression models with probit and logit links to three dimensional panel data. The package
extends R package mvord with a further multiple measurement object and a further correlation
structure, to accommodate for multiple measurements which exhibit cross-sectional and longitudinal
dependence. Medium-term the functionalities will be incorporated in package mvord. As in
mvord, regression coefficients and threshold parameters are varying across the multiple response
dimensions in the default implementation. However, constraints can be defined by the user if a
reduction of the parameter space is desired.
}
\details{
see \code{\link{mvordflex}}
}
